requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.6
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-classy-tags==0.8.0
django-extensions==1.9.6
django-geojson==2.10.0
django-leaflet==0.22.0
django-test-without-migrations==0.6
geopy==1.11.0
lxml==4.0.0
psycopg2==2.7.3.1
python-decouple==3.1
pytz==2017.2
six==1.11.0
static3==0.7.0
